add get devices
